requirements.txt
# core
numpy~=1.24.3
pandas
scikit-learn
pyyaml~=6.0.2

# nlp
hazm
nltk
stanza
beautifulsoup4
lxml
transformers~=4.56.2
tokenizers~=0.22.1
sentencepiece~=0.2.1

# torch (installed in Dockerfile to control cpu/gpu variant)
# torch

torch~=2.8.0